+++ src/InDesign/Command/Table.php
@@ -14,6 +14,7 @@
 namespace Mds\PimPrint\CoreBundle\InDesign\Command;
 
 use League\Flysystem\FilesystemException;
+use Mds\PimPrint\CoreBundle\InDesign\Command\Traits\DefaultLocalizedTrait;
 use Mds\PimPrint\CoreBundle\InDesign\Command\Traits\FitTrait;
 use Mds\PimPrint\CoreBundle\InDesign\Command\Traits\ImageCollectorTrait;
 use Mds\PimPrint\CoreBundle\InDesign\Text;
@@ -28,6 +29,7 @@ class Table extends AbstractBox implements ImageCollectorInterface, ComponentInt
 {
     use FitTrait;
     use ImageCollectorTrait;
+    use DefaultLocalizedTrait;
 
     /**
      * Command name.